+++ FrontEnd/App.tsx
@@ -1,32 +1,14 @@
 import React from 'react';
-
-import {StyleSheet, Text, View} from 'react-native';
+import Open from './android/app/src/Screens/Registers/Open';
+import {View} from 'react-native';
 
 function App(): React.JSX.Element {
   return (
     <View>
-      <Text>Hello</Text>
+     <Open/>
     </View>
   );
 }
 
-const styles = StyleSheet.create({
-  sectionContainer: {
-    marginTop: 32,
-    paddingHorizontal: 24,
-  },
-  sectionTitle: {
-    fontSize: 24,
-    fontWeight: '600',
-  },
-  sectionDescription: {
-    marginTop: 8,
-    fontSize: 18,
-    fontWeight: '400',
-  },
-  highlight: {
-    fontWeight: '700',
-  },
-});
 
 export default App;
