+++ FrontEnd/android/app/src/Screens/Registers/Open.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+function App(): React.JSX.Element {
+  return (
+    <View style={styles.sectionContainer}>
+      <Text style={styles.sectionTitle}>Hello</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  sectionContainer: {
+    marginTop: 32,
+    paddingHorizontal: 24,
+    color:'red',
+    backgroundColor:''
+  },
+  sectionTitle: {
+    fontSize: 24,
+    fontWeight: '600',
+  },
+  sectionDescription: {
+    marginTop: 8,
+    fontSize: 18,
+    fontWeight: '400',
+  },
+  highlight: {
+    fontWeight: '700',
+  },
+});
+
+export default App;
